requirements.txt
gradio
gradio_leaderboard
pandas
PyGithub